Split tests using real and emulator firestore
