+++ __tests__/real/new_user_login.test.ts
@@ -1,6 +1,6 @@
 import firebase, { firestore } from 'firebase'
-import config from '../firebase_config.json'
-import { generateRandomUserConfig } from '../src/test_util'
+import config from '../../firebase_config.json'
+import { generateRandomUserConfig } from '../../src/test_util'
 
 // 元々存在しているユーザーがauthで認証でき、ログインに相当する自分のFirestoreの変更ができるかのテスト
 // テスト用の新しいユーザーがサインアップ、テスト後にデータを削除する
